Set `dirty` on ComplexStrings to detect recursion
Remove the (incomplete) implementation of locals.__stack__.  The stack
would need to be passed around from one call to another to stay
up-to-date.  This would require changes in how expressions are
represented (regular objects which store state instead of functions) or
changes to what these functions return (return value and the new stack
instead of the return value alone).

Instead, set a `dirty` flag on ComplexStrings and make use of the
closure to change it from within the expression function.