ComplexString's dirty should not be shared by all complexStrings
When dirty was defined as a variable in the ComplexString scope, it was
avilable and effectively shared by all complexStrings.  This change
moves into the return function so that it's only available to one
complexString, and moves all the expression logic into a closure so
that dirty's state persists across mutlitple calls to the same
complexString.