Remove _step
The _step function was called every time there was a need to convert
an expression's [locals, memberExpr]-typed return value into a
{locals: locals, self: memberExpr}-typed one.

Removed _step in favor of a destructing assignment-like syntax.  The
actual destructing assignment (with its benefits) is part of JS1.8
which is a superset of features over ES5.  V8 only supports ES5, so
it's not possible to do `var [locales, memberExpr] = expr(...)` and be
cross-engine compatible at the same time.