Change how variables are stored in locals (+tests)
Macro arguments are no longer resolved in the callExpression.  This
made it impossible to pass an entity or a hash to a macro, and resulted
in weird behavior when passing a macro.

Instead, argument expressions are now simply called once, so that the
actual arguments passed to the macro are objects, not expressions, and
they have their own locals stored in the macro's locals, too.

For instance, in a callExpression like this one:

  plural(7)

'7' is a stringLiteral expression, and calling it returns 7, which is
then passed to the macro expression.

Similarly, for a macro call like so:

  getDefaultValue(brandName)

'brandName' is an indentifier expression, and calling it return the
entity object called 'brandName.'

Note that we must call the expression before passing it to the macro
expression.  Otherwise, we would be passing expression functions around
(e.g., propertyExpressions) instead of actual objects that they
evaluate to, and then we'd call these expression functions with wrong
locals.