Remove logicalExpressionLeft
The old JS parser would always produce a LogicalExpression in the
`test` symbol of a ConditionalExpression.  For instance, `n == 1` in
a ConditionalExpression would be parsed as ConditionalExpression ->
LogicalExpression -> BinaryExpression.  This required
LogicalExpressions to handle the case where they don't have the
operator nor the RHS.

The new parser skips the LogicalExpression in above cases.  Hence,
there's no more need for the special-case LHS-only handling in
LogicalExpressions.

This commit also adds a few tests for ConditionalExpressions.  One
thing of note:  UnaryExpressions, BinaryExpressions and
LogicalExpressions currently _resolve their components to primitive
values prior to doing the operation of the expression.  I.e., arrays
and hashes are resolved to their default values first.  This has
important consequences for testing truthiness--an empty array [] and an
empty hash {}, although truthy in JavaScript, are first resolved to
their default values, which right now are `undefined`.  This means that
they both are falsy in L20n.