Fix duplicate imports
The terminal callback (preprocess for ProcessedResource, compile for
Context) executes when the object in question `isComplete`.  The
completion test used to consist of checking if all children were ready.

There was a problem with this approach, though.  The list of children
was only known after all importResource were executed.  importResource
is alaso the very same place where the terminal callback is passed.

If all the importResource happened (or were specified to be)
synchronous, isComplete would return true for the first child resource,
because at the time it was checking, other importResources haven't run
yet.

This commit fixes this by also counting the import statements in the
AST and making sure that isComplete returns true only if the number of
imported child resources (`totalResources`) is equal to the length of
the `importedResources` array.