Split normalizeURL into two functions
relativeToSelf takes the URL and makes it relative to the parent
resource.  E.g., if locales/pl/foo.lol imports bar.lol, the produced
URL will be locales/pl/bar.lol.

normalizeURL removes the unnecessary parts of the URL.  E.g.,
locales/pl/./common/../foo.lol becomes locales/pl/foo.lol.

relativeToSelf is only used when the URI does not contain the l10n:
schema so that the imported resource can be properly located from the
location of the script execution.

normalizeURL is called for all URLs so that duplicate imports of the
same resource (but from different places) can be detected and
prevented.