Bug 883636 - Create CLI tools. r=gandalf
Create CLI tools for parsing, compiling and perf measurements.

`parse.js` is a utility script that uses L20n's Parser to parse a string into
an AST and output the result.

`compile.js` is a utility script that uses L20n's Compiler to parse and compile
a string into Entity objects and output the result.

`perf` is a utility script to measure the performance of parsing, compilation
and getting entities.