Remove the Babel Polyfill requirement
The `for..of` syntax requires iterators which are only supported by Babel with
a special polyfill which emulates a full ES6 environment.  The polyfill is
quite hefty weighing at 25KB.  By replacing `for..of` with a different method
of iterating over an array we can remove this requirement and make l20n.js
more lightweight in the webcompat version.

http://babeljs.io/docs/usage/polyfill/

Additionally, I suggest that we turn on Babel's 'loose' transpilation.  From
their page:

    Loose mode enables certain transformers to generate cleaner output that
    lacks specific ES6 edgecases. These edgecases are either unlikely to
    appear in your code or the inclusion of them introduces significant
    overhead.

    As a result of loose mode code will execute faster and be much more
    readable and comparable to the original but will deviate from the spec
    in slight ways.

https://babeljs.io/docs/advanced/loose/

Lastly, Babel also includes some optional transformers which make the code
cleaner.  Let's turn them on as well.

http://babeljs.io/docs/advanced/transformers/#optional