Bug 1206656 - Remove supportedlanguageschange event
We don't currently have a good use-case for this event and I think it'd be
safer if we removed it for now.  We're also not sure if it should be emitted by
the document or some other object.

If a need emerges in the future, we can re-add this event.  But first, I think
we should figure out where it fits in the NGA and which object is responsible
for emitting it.

In this commit I'm also removing document.localized which isn't very useful
right now (document.l10n.ready should be used instead).