Transpile classes for non-nightly Gecko
`grunt compat` now takes a second argument which can be `compat` or `gecko`
and which controls the set of babel transformers to be used for transpilation.

The default is to use the `compat` transformers set, so the following commands
are the same:

    grunt compat:web
    grunt compat:web:compat

The `gecko` variant currently has only one transformer enabled: `es6.classes`
which is need for non-nighly Gecko browsers.

    grunt build
    grunt compat:web:gecko
    open examples/webgecko.html

Additionally, `grunt gaia` now does the right thing and copies the correct
version of l20n.js.