Add MessageContext.formatToPrimitive
This is a bit like formatToParts() on other Intl formatters.  Regular format()
always returns a string, even if the entity doesn't have a value.  Sometimes
it's useful to actually know that there was no value, e.g. when the translation
is only for HTML attributes and shouldn't touch the textContent.

For instance:

  input =
    [placeholder] Hello

The resolver will see that there is no value and will try to find a default
trait.  Since there isn't any it will return FTLNone.

The context can then do one of two things:

 - format(): call toString() on the FTLNone, which produces a string, or
 - formatToPrimitive(): return null for the FTLNone.