Search for localizations in document.l10n only
Looking up <localization> elements poses a few problems:

  - we don't have them in l20n-chrome-html,
  - they can't have ids because they might be part of the anonymous content
    bound by XBL more than once,
  - some bindings currently use <stringbundles> defined in different bindings
    and they refer to them from the XBL code.

Re. the last point, here's an example:  tabbrowser-tabs uses the
tabs-empty-tab-title string which comes from the <stringbundle>/<localization>
defined in the tabbrowser binding.  Furthermore, the <tab> which uses the
binding is not part of the anonymous content and doesn't have easy access to
tabbrowser's anonymous content (other than hardcoding it).