Bug 1317989 - Isolate all interpolations in MessageContext. r=gandalf
When a pattern is resolved, all interpolations will be by default wrapped in
the Unicode Isolation Control Codes: FSI (U+2068) and PDI (U+2069).  These
control codes are used by the Unicode Bidi Algorithm as hints to correctly render weak
characters which could be affected by the directionality run of adjacent strong
characters.

Given the following FTL:

    foo = Foo
    bar = { foo } Bar

`bar` will format as:

    FSI + "Foo" + PDI + " Bar"

See `test/intl/isolating_test.js` for more examples.

This behavior is on by default and can be controlled via the `useIsolating`
option passed to the `MessageContext` constructor.  Turning it off might be
desirable in environments which don't support Unicode well, as well as for
tools and tests.

    const ctx = new MessageContext(lang, {
        useIsolating: false
    });

The FSI and PDI control codes were introduced in Unicode 6.3.  More information
about them can be found in the folowing W3C recommendation:

    https://www.w3.org/International/articles/inline-bidi-markup/#nomarkup