Add spans and annotations to Entries
This change allows consumers of the fluent-syntax API to display nice error
messages.

`Entry` instances now have a `span: { from, to }` field. `from` and
`to` are positions relative to the start of the file.

```json
{
    "type": "Message",
	"annotations": [],
	"id": {
	    "type": "Identifier",
	    "name": "hello-world"
	},
	"value": {
	    "type": "Pattern",
	    "elements": [
	    {
		"type": "StringExpression",
		"value": "Hello, world!"
	    }
	    ],
	    "quoted": false
	},
	"attributes": null,
	"comment": null,
	"span": {
	    "from": 44,
	    "to": 71
	}
},
```

Parsing errors are now collected into the `annotations` field of `Entry`
instances.  In the future, we can add warnings and notes there, too.

```json
{
    "type": "Junk",
    "annotations": [
    {
	"message": "Missing field",
	"name": "ParseError",
	"pos": 85
    }
    ],
    "content": "hello-world Hello, world!\n\n",
    "span": {
	"from": 73,
	"to": 100
    }
},
```

As per [Error-reporting][], in the future we may considering adding `labels` to
annotations to provide explanation of the error and suggested ways of fixing
it.

[Error-reporting]: https://github.com/projectfluent/fluent/wiki/Error-reporting

`fluent-syntax` now also exposes two helpers: `lineOffset(source, pos)` and
`columnOffset(source, pos)`, both returning 0-based offset numbers.  They may
be used to place gutter markers in editors or to calculate the number of
context lines to show around an error.

Lastly, the `tools/parse` script now prints the following output for errors.
I didn't use any colors at this point because I didn't want to make the patch
too complicated.

    ! SyntaxError on line 3:
      | @xxx
      … ^----- Expected entry

    ! SyntaxError on line 5:
      | hello-world Hello, world!
      …             ^----- Missing field

    ! SyntaxError on line 7:
      | shared-photos = { $user_name { $photo_count ->
      …                              ^----- Expected token }
      |     [0] hasn't added any photos yet
      |     [one] added a new photo
      |    *[other] added { $photo_count } new photos
      | }.

    ! SyntaxError on line 17:
      | liked-comment = { $user_name } liked your comment on { $user_gender ->
      |     [male] his
      |     [female] her
      |     [other] their
      … ^----- Missing default variant
      | } post.