Add CHANGELOG files to all packages
Important changes should include an edit to the `CHANGELOG.md` file in the
_Unreleased_ section.

At the time of the release, `make CHANGELOG.md SINCE=x.y.z` can be used to
insert the `git log` output since the version `x.y.z` at the beginning of
`CHANGELOG.md`. Then, edit the changelog by hand:

  - Rename the _Unreleased_ section to reflect the version that's being
    released.

  - Create a new _Unreleased_ section at the top of the file.

  - Edit the lits of commits to be human-readable: remove noisy commits and
    change descriptions and ordering if needed.

  - Highlight removals and deprecations.  Give a short summary of steps
    required to upgrade.

When `CHANGELOG.md` is ready, commit it together with the updated
`package.json`.  Tag the commit using the `name@x.y.z` naming scheme and put
the list of changes from `CHANGELOG.md` in the tag body.