Introduce behavior and structure testing
    Behavior Tests

FTL files which may include special markup with the information about the
expected error. The test runner uses a preprocessor to parse and remove the
error annotation from the file. If there is no error annotation, the file is
expected to parse without errors.

    Structure Tests

FTL + JSON AST pairs. JSON files are generated via `make fixtures` and reviewed
by a human. The test runner compares the output of the parser against the saved
AST.