fluent-react: Introduce the withLocalization HOC.
The withLocalization decorator is a higher-order component which can be
used to gain imperative access to a Localization.

Sometimes it's useful to imperatively format a translation rather than
rely on the declarative `<Localized />` API.  Good examples of this are
the `window.alert` and `window.confirm` functions.

The withLocalization HOC connects a component to its enclosing
LocalizationProvider and injects the formatString prop into the wrapped
component.  The formatString method returns a translation given an id
and an optional arguments object.

    import { withLocalization } from 'fluent-react/compat';

    function HelloButton(props) {
        const { formatString } = props;

        return (
            <button onClick={() => alert(formatString('hello'))}>
                👋
            </button>
        );
    }

    export default withLocalization(HelloButton);