Forbid exclusive tests
Forgetting to remove the `.only` method on a test can result in
a reduced test coverage.  Introduce a minimal ESLint config for tests
which checks for exclusive tests. No other linting rules are enabled for
now because our test files haven't been linted in the past and there's
a lot to fix.