Require = after the id (but not really) (#132)
In Fluent Syntax 0.5, the identifier must be followed by an equals sign (=)
even if the message doesn't have a value.

    key =
        .attribute = Attribute Value

For some time fluent.js will allow both the new (0.5) and the old 0.4 syntax,
which means that the following will still parse but the Serializer will insert
the = after the key.

    key
        .attribute = Attribute Value

This patch also fixes Bug 1406880 - Decide where multiline Pattern spans should
start. The patterns now start when their first

    key =
        Value
        ^---- Pattern span start

Lastly, this patch forbids null variant and attribute values in the tooling
parser. Empty values should be written with the {""} idiom:

    key =
        {
           *[valid] {""}
            [invalid]
        }

The runtime parser parses them as {val: null} for performance reasons.