fluent-react: Filter props with <Localized attrs={{…}}> (#141)
The <Localized> component now requires the `attrs` prop to set any localized
attributes as props on the wrapped component. `attrs` should be an object with
attribute names as keys and booleans as values.

```jsx
<Localized id="type-name" attrs={{placeholder: true}}>
    <input
        type="text"
        placeholder="Localizable placeholder"
        value={name}
        onChange={…}
    />
</Localized>
```

By default, if `attrs` is not passed, no attributes will be set. This is
a breaking change compared to the previous behavior: in `fluent-react`
0.4.1 and before `<Localized>` would set _all_ attributes found in the
translation.