Inline Patterns may start with any character. (#150)
`}`, `.`, `*` and `[` are only special when they appear at the beginning of indented Pattern lines. When a Pattern starts on the same line as `id =` or `[variant key]`, its first character doesn't carry any special meaning and it may be one of those four ones as well.

See also https://bugzilla.mozilla.org/show_bug.cgi?id=1436683.

The tests for this uncovered a bug in how the runtime parser treated null attributes and variants. For compliance with fluent-syntax these are syntax errors now.