Relax the engines support in package.json for node6 support.
This blocks to mozilla/addons-linter#1789 (and thus the langpack signing
for Firefox 60).

fluent-syntax is compatible with node 6 when using the compat build but
still can't be installed given that it requires node 8+ in the
packages.json.

Can we relax that a little bit?

Addons-linter is a dependency of `web-ext` which is being installed on
end-users machines and thus has a bit more LTS requirements and has to
support node6 while it's still LTS.