fluent-react: Protect void elements against translated text content
Ignore the text content found in the translated markup when the element passed
as a prop to <Localized> is a known void element. This prevents the "element is
a void element tag and must neither have children nor use
dangerouslySetInnerHTML" error in React.

For instance, <input> is a known void element and it must not have the text
content set by the translation in the following example:

    <Localized id="hello" text-input={<input type="text" />}>
        <div />
    </Localized>

    hello = Hello, <text-input>invalid text content</text-input>.

Fix #174.