Parse Comments as standalone if they're attached to Junk
Syntax 0.6 changes the parsing logic of Comments. Comments attached to what
ends up being Junk are not part of Junk anymore. Instead, they end up as
standalone Comments in the final Resource.

This PR also changes the behavior of FluentParser.parseEntry() to ignore all
valid comments and only start parsing when a Message or a Term is encountered.
Comments with syntax errors are not skipped, however.